Added master/slave board relationship
This is the second revision of the code & hardware. The sound board now acts as a master and the LED board is a slave. All buttons are routed solely to the sound board, and it sends signals to the LED board via hardware serial. There is also a pin expander IC incorporated instead of the pins on the underside of the Teensy 3.1.